+++ CHANGELOG.md
@@ -64,6 +64,9 @@
 
 ### Snowpark pandas API Updates
 
+#### New Features
+- Added support for the `dtypes` parameter of `pd.get_dummies`
+
 #### Improvements
 
 - Improved performance of `Series.to_snowflake` and `pd.to_snowflake(series)` for large data by uploading data via a parquet file. You can control the dataset size at which Snowpark pandas switches to parquet with the variable `modin.config.PandasToSnowflakeParquetThresholdBytes`.